requirements.txt
# Wafer Detection Agent - Requirements
# =====================================

# Core numerical computation
numpy>=1.24

# Image handling (resize synthetic wafers)
Pillow>=9.5

# PyTorch (model + inference)
# For CPU-only installation, use: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0
torchvision>=0.15

# TensorFlow (for ViT and .h5 models)
tensorflow>=2.15.0

# Data visualization (optional but useful for debugging)
matplotlib>=3.7

# Progress bars for long operations
tqdm>=4.65

# Data manipulation
pandas>=2.0

# Computer vision for wafer mask extraction
opencv-python>=4.8

# =====================================
# Additional dependencies for the Agent system
# =====================================

# None required - the google.adk mock is included in the project
# For production with real Google ADK, you would add:
# google-adk>=0.1.0  # (hypothetical package)

# =====================================
# API Server dependencies
# =====================================
fastapi>=0.104
uvicorn>=0.24
python-multipart>=0.0.6

# =====================================
# Database & Analytics dependencies
# =====================================
sqlalchemy>=2.0.23
pydantic-settings>=2.0.0
reportlab>=4.0.7
openpyxl>=3.1.0
